move ssl stuff out of auth folder, rename redirect url to callback url, add a new redirect url, move config to a separate object inside the auth package
